security: prevent LFI if the disk adapter is being used
Hopefully this will prevent it. Have to dig deeper into this by going through som SecLists for LFI. Before this fix I was able to delete files on the filesystem such as /etc/passwd if the user was a root user. Was not able to view any other file than json files which had the same content structured as the secret files, though.

+++ src/server/services/disk.js
@@ -3,6 +3,7 @@ const { nanoid } = require('nanoid');
 const config = require('config');
 
 const { encrypt, decrypt } = require('../helpers/crypto');
+const isLFIAttempt = require('../helpers/lfi');
 
 async function upload(encryptionKey, fileUpload) {
     const filename = nanoid();
@@ -24,6 +25,10 @@ async function upload(encryptionKey, fileUpload) {
 }
 
 async function download(key, encryptionKey) {
+    if (isLFIAttempt(key)) {
+        return 'LFI attempt';
+    }
+
     try {
         const data = await fs.readFile(key, 'utf-8');
 
@@ -38,6 +43,10 @@ async function download(key, encryptionKey) {
 }
 
 async function remove(key) {
+    if (isLFIAttempt(key)) {
+        return 'LFI attempt';
+    }
+
     const data = await fs.unlink(key);
 
     return data;
